Now load scraper iframe in background.html.
Before we were just loading the iframe in the options page, but this
could have resulted in the user closing it before the scraping finished
and other fun bugs. Also the user would have seen the loading symbol
etc. This way we completely hide the iframe away in the background page.