Refactor replacing indexeddb with chrome.storage.
With indexeddb we cant use the same database in both content script
files and extension files (background.html etc). We really want to be
able to use backbone models in both and we want them to both be using
the same data source.

So we have switched to using chrome.storage. Both content scripts and
extension scripts can access the same chrome.storage sharing data
between both. To use this with backbone we have added in
backbone.chromestorage which overrides backbone.sync to work with chrome
storage.